+++ fairmq/shmem/Socket.h
@@ -155,7 +155,7 @@ class Socket final : public fair::mq::Socket
         }
     }
 
-    int Send(MessagePtr& msg, const int timeout = -1) override
+    int64_t Send(MessagePtr& msg, const int timeout = -1) override
     {
         int flags = 0;
         if (timeout == 0) {
@@ -191,7 +191,7 @@ class Socket final : public fair::mq::Socket
         return static_cast<int>(TransferResult::error);
     }
 
-    int Receive(MessagePtr& msg, const int timeout = -1) override
+    int64_t Receive(MessagePtr& msg, const int timeout = -1) override
     {
         int flags = 0;
         if (timeout == 0) {
